Stop listkeys test from caring about whether public keys have CKA_PRIVATE
If we *choose* to depend on the tools in the build environment by building
the SoftHSM database as part of the test process instead of having a
pre-packaged one in the repository... and if we *choose* to use pkcs11-tool
to create objects instead of using softhsm2-util or p11tool which would
let us explicitly manage the CKA_PRIVATE state of the objects we create...
then let's log in to the token before enumerating the public keys.

Otherwise we are *deliberately* making the test fragile because it depends
on behaviour we *know* has changed in the environment quite recently. We're
supposed to be testing libp11, not the build environment.